Created Brand model for linking famous brands on a Market Created new geography app for storing Country/Region information Wrote migrations to copy across the Country/Region data from markets
