+++ geography/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from .models import Region, Country
+
+
+admin.site.register(Region)
+admin.site.register(Country)
