+++ geography/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class GeographyConfig(AppConfig):
+    name = 'geography'
+    verbose_name = 'Geography'
