+++ navigator/settings/base.py
@@ -36,6 +36,7 @@ INSTALLED_APPS = [
     'core',
     'contact',
     'markets',
+    'geography',
 ]
 
 MIDDLEWARE_CLASSES = [
